+++ packages/snap-preact/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.77.0](https://github.com/searchspring/snap/compare/v0.76.1...v0.77.0) (2026-04-24)
+
+### Bug Fixes
+
+- **preact/recommendationinstantiator:** ensuring controllers are being re-used when possible ([712ef43](https://github.com/searchspring/snap/commit/712ef438fe8a209363a92c8f9ce872fddabc3118))
+- **preact/recommendationinstantiator:** refactoring removing await hack and reusing controller ([26fd6d5](https://github.com/searchspring/snap/commit/26fd6d5070d0c3db0beba3424cf1de59b8865e4f))
+- **preact/recommendationinstantiator:** supporting SPA behvior for targeting ([e9f0f48](https://github.com/searchspring/snap/commit/e9f0f48216a04ae7abbedb43e4f1a1b39465ad47))
+
 ## [0.76.1](https://github.com/searchspring/snap/compare/v0.76.0...v0.76.1) (2026-04-08)
 
 **Note:** Version bump only for package @searchspring/snap-preact